Move Selector to model
